The "Bones merging script" : Addon edition
Cleaning up 3D models for VRChat can take a surpsingly
high amount of time when dealing with useless bones.
There's no way to merge bones AND their associated weight
painting in Blender, since the weight paints are associated
with the vertex groups, and there's no easy way to merge
them together.

There are modifiers to merge vertex groups weights painting,
but these take an insane amount of time to setup correctly.

So I made this addon. Just select all the bones to merge,
ensure the merge destination is active, then "Right-click >
Merge with active" and you're done !

It's as simple as that. I used my tool on various avatars
already, with perfect results.

Signed-off-by: Voyage <voyage@miouyouyou.fr>