Multiple fixes
* Bail out if X mirror mode is activated

Turns out that you can mirror actions dones on part of
the mesh, or the armature, through a special operator.
This is one is different from the "Mirror" modifier, but
still acts in a similar fashion.

This leads to issues when trying to merge bones on legs
or arms. The mirror operator will try to mimic this on
the other part, but will completely forgot about the
vertex groups.

* Added a Japanese translation

Along with a translation for the error messages.

* Named the whole thing "Voyage Bones Merger"

Still, most people won't understand the 'Myy' reference.

Signed-off-by: Voyage <voyage@miouyouyou.fr>