Add the active bone vertex group if needed
When checking the first time, if no mesh contains
the active bone vertex group name, we'll stop there.

But if any of them contains it, and others don't,
we'll just add it to all of them.

The reason being, if we ACTUALLY do the 'merging' and
remove the bones afterward, while skipping any mesh
that doesn't have the target vertex group, we run the
risk of removing bones without merging any weight paint
at all, and thus losing these bones AND their weight paints.

Signed-off-by: Voyage <voyage@miouyouyou.fr>